Refactored tests into categories
