+++ application/tests/tests_application/__init__.py
@@ -1,0 +1,2 @@
+from .test_functional import *
+from .test_login_redirect_helper import *
