+++ application/tests/tests_application/test_functional.py
@@ -10,8 +10,7 @@ from django.test import Client, TestCase
 
 from timeline_logger.models import TimelineLog
 
-from application.middleware import CustomAuthenticationHandler
-from application.models import Application, UserDetails, ApplicantName, ApplicantPersonalDetails, ApplicantHomeAddress, \
+from ...models import Application, UserDetails, ApplicantName, ApplicantPersonalDetails, ApplicantHomeAddress, \
     ChildcareType, Reference, AdultInHome, Arc, ChildInHome, CriminalRecordCheck, EYFS
