+++ application/tests/tests_business_logic/__init__.py
@@ -1,0 +1,8 @@
+from .test_childcare_type_logic import *
+from .test_dbs_check_logic import *
+from .test_EYFS_login import *
+from .test_first_aid_training_logic import *
+from .test_health_logic import *
+from .test_people_in_your_home_logic import *
+from .test_personal_logic import *
+from .test_reference_logic import *
