+++ application/tests/tests_selenium/__init__.py
@@ -1,0 +1,3 @@
+from .apply_as_a_childminder import *
+from .test_eyfs import *
+from .selenium_task_executor import *
