+++ application/tests/tests_views/test_account_selection.py
@@ -1,0 +1,8 @@
+from .test_views import *
+
+
+class AccountSelectionTest(ViewsTest):
+
+    def test_url_resolves_to_page(self):
+        found = resolve(settings.URL_PREFIX + '/sign-in/new-application/')
+        self.assertEqual(found.func, account_selection)
