+++ application/tests/tests_views/test_existing_application.py
@@ -1,0 +1,8 @@
+from .test_views import *
+
+
+class ExistingApplicationTest(ViewsTest):
+
+    def test_url_resolves_to_page(self):
+        found = resolve(settings.URL_PREFIX + '/sign-in/')
+        self.assertEqual(found.func, existing_email)
