+++ application/tests/tests_views/test_start_page.py
@@ -1,0 +1,8 @@
+from .test_views import *
+
+
+class StartPageTest(ViewsTest):
+
+    def test_root_url_resolves_to_start_page_view(self):
+        found = resolve(settings.URL_PREFIX + '/')
+        self.assertEqual(found.func, start_page)
