+++ application/tests/tests_views/test_task_status.py
@@ -1,0 +1,327 @@
+from .test_views import *
+
+
+class TaskStatusTest(ViewsTest):
+
+    def test_status_update_with_tasks_not_started(self):
+        test_application_id = 'f8c42666-1367-4878-92e2-1cee6ebcb48c'
+        test_login_id = '004551ca-21fa-4dbe-9095-0384e73b3cbe'
+        application = models.Application.objects.create(
+            application_id=(UUID(test_application_id)),
+            application_type='CHILDMINDER',
+            application_status='DRAFTING',
+            cygnum_urn='',
+            login_details_status='COMPLETED',
+            personal_details_status='COMPLETED',
+            childcare_type_status='COMPLETED',
+            first_aid_training_status='COMPLETED',
+            eyfs_training_status='COMPLETED',
+            criminal_record_check_status='COMPLETED',
+            health_status='COMPLETED',
+            references_status='COMPLETED',
+            people_in_home_status='NOT_STARTED',
+            declarations_status='NOT_STARTED',
+            date_created=datetime.datetime.today(),
+            date_updated=datetime.datetime.today(),
+            date_accepted=None,
+            order_code=None
+        )
+        user = models.UserDetails.objects.create(
+            application_id=application,
+            login_id=(UUID(test_login_id)),
+            email='',
+            mobile_number='',
+            add_phone_number='',
+            email_expiry_date=None,
+            sms_expiry_date=None,
+            magic_link_email='',
+            magic_link_sms=''
+        )
+
+        application_status_context = dict({
+            'application_id': test_application_id,
+            'login_details_status': application.login_details_status,
+            'personal_details_status': application.personal_details_status,
+            'childcare_type_status': application.childcare_type_status,
+            'first_aid_training_status': application.first_aid_training_status,
+            'eyfs_training_status': application.eyfs_training_status,
+            'criminal_record_check_status': application.criminal_record_check_status,
+            'health_status': application.health_status,
+            'reference_status': application.references_status,
+            'people_in_home_status': application.people_in_home_status,
+            'declaration_status': application.declarations_status,
+            'all_complete': False,
+            'confirm_details': False
+        })
+        temp_context = application_status_context
+        del temp_context['declaration_status']
+        assert (('NOT_STARTED' in temp_context.values()) == True)
+
+    def test_status_update_with_tasks_in_progress(self):
+        test_application_id = 'f8c42666-1367-4878-92e2-1cee6ebcb48c'
+        test_login_id = '004551ca-21fa-4dbe-9095-0384e73b3cbe'
+
+        application = models.Application.objects.create(
+            application_id=(UUID(test_application_id)),
+            application_type='CHILDMINDER',
+            application_status='DRAFTING',
+            cygnum_urn='',
+            login_details_status='COMPLETED',
+            personal_details_status='COMPLETED',
+            childcare_type_status='COMPLETED',
+            first_aid_training_status='COMPLETED',
+            eyfs_training_status='COMPLETED',
+            criminal_record_check_status='COMPLETED',
+            health_status='COMPLETED',
+            references_status='COMPLETED',
+            people_in_home_status='IN_PROGRESS',
+            declarations_status='NOT_STARTED',
+            date_created=datetime.datetime.today(),
+            date_updated=datetime.datetime.today(),
+            date_accepted=None,
+            order_code=None
+        )
+        user = models.UserDetails.objects.create(
+            application_id=application,
+            login_id=(UUID(test_login_id)),
+            email='',
+            mobile_number='',
+            add_phone_number='',
+            email_expiry_date=None,
+            sms_expiry_date=None,
+            magic_link_email='',
+            magic_link_sms=''
+        )
+        application_status_context = dict({
+            'application_id': test_application_id,
+            'login_details_status': application.login_details_status,
+            'personal_details_status': application.personal_details_status,
+            'childcare_type_status': application.childcare_type_status,
+            'first_aid_training_status': application.first_aid_training_status,
+            'eyfs_training_status': application.eyfs_training_status,
+            'criminal_record_check_status': application.criminal_record_check_status,
+            'health_status': application.health_status,
+            'reference_status': application.references_status,
+            'people_in_home_status': application.people_in_home_status,
+            'declaration_status': application.declarations_status,
+            'all_complete': False,
+            'confirm_details': False
+        })
+        temp_context = application_status_context
+        del temp_context['declaration_status']
+        assert (('IN_PROGRESS' in temp_context.values()) == True)
+
+    def test_status_update_with_tasks_completed_except_declarations(self):
+        test_application_id = 'f8c42666-1367-4878-92e2-1cee6ebcb48c'
+        test_login_id = '004551ca-21fa-4dbe-9095-0384e73b3cbe'
+
+        application = models.Application.objects.create(
+            application_id=(UUID(test_application_id)),
+            application_type='CHILDMINDER',
+            application_status='DRAFTING',
+            cygnum_urn='',
+            login_details_status='COMPLETED',
+            personal_details_status='COMPLETED',
+            childcare_type_status='COMPLETED',
+            first_aid_training_status='COMPLETED',
+            eyfs_training_status='COMPLETED',
+            criminal_record_check_status='COMPLETED',
+            health_status='COMPLETED',
+            references_status='COMPLETED',
+            people_in_home_status='COMPLETED',
+            declarations_status='NOT_STARTED',
+            date_created=datetime.datetime.today(),
+            date_updated=datetime.datetime.today(),
+            date_accepted=None,
+            order_code=None
+        )
+        user = models.UserDetails.objects.create(
+            application_id=application,
+            login_id=(UUID(test_login_id)),
+            email='',
+            mobile_number='',
+            add_phone_number='',
+            email_expiry_date=None,
+            sms_expiry_date=None,
+            magic_link_email='',
+            magic_link_sms=''
+        )
+        application_status_context = dict({
+            'application_id': test_application_id,
+            'login_details_status': application.login_details_status,
+            'personal_details_status': application.personal_details_status,
+            'childcare_type_status': application.childcare_type_status,
+            'first_aid_training_status': application.first_aid_training_status,
+            'eyfs_training_status': application.eyfs_training_status,
+            'criminal_record_check_status': application.criminal_record_check_status,
+            'health_status': application.health_status,
+            'reference_status': application.references_status,
+            'people_in_home_status': application.people_in_home_status,
+            'declaration_status': application.declarations_status,
+            'all_complete': False,
+            'confirm_details': False
+        })
+        temp_context = application_status_context
+        del temp_context['declaration_status']
+        assert (('IN_PROGRESS' in temp_context.values()) == False)
+        assert (('NOT_STARTED' in temp_context.values()) == False)
+
+    def test_status_update_with_tasks_completed_with_declarations_to_do(self):
+        test_application_id = 'f8c42666-1367-4878-92e2-1cee6ebcb48c'
+        test_login_id = '004551ca-21fa-4dbe-9095-0384e73b3cbe'
+
+        application = models.Application.objects.create(
+            application_id=(UUID(test_application_id)),
+            application_type='CHILDMINDER',
+            application_status='DRAFTING',
+            cygnum_urn='',
+            login_details_status='COMPLETED',
+            personal_details_status='COMPLETED',
+            childcare_type_status='COMPLETED',
+            first_aid_training_status='COMPLETED',
+            eyfs_training_status='COMPLETED',
+            criminal_record_check_status='COMPLETED',
+            health_status='COMPLETED',
+            references_status='COMPLETED',
+            people_in_home_status='COMPLETED',
+            declarations_status='NOT_STARTED',
+            date_created=datetime.datetime.today(),
+            date_updated=datetime.datetime.today(),
+            date_accepted=None,
+            order_code=None
+        )
+        user = models.UserDetails.objects.create(
+            application_id=application,
+            login_id=(UUID(test_login_id)),
+            email='',
+            mobile_number='',
+            add_phone_number='',
+            email_expiry_date=None,
+            sms_expiry_date=None,
+            magic_link_email='',
+            magic_link_sms=''
+        )
+        application_status_context = dict({
+            'application_id': test_application_id,
+            'login_details_status': application.login_details_status,
+            'personal_details_status': application.personal_details_status,
+            'childcare_type_status': application.childcare_type_status,
+            'first_aid_training_status': application.first_aid_training_status,
+            'eyfs_training_status': application.eyfs_training_status,
+            'criminal_record_check_status': application.criminal_record_check_status,
+            'health_status': application.health_status,
+            'reference_status': application.references_status,
+            'people_in_home_status': application.people_in_home_status,
+            'declaration_status': application.declarations_status,
+            'all_complete': True,
+            'confirm_details': False
+        })
+        assert (application_status_context['declaration_status'] == 'NOT_STARTED')
+
+    def test_status_update_with_tasks_completed_with_declarations_in_progress(self):
+        test_application_id = 'f8c42666-1367-4878-92e2-1cee6ebcb48c'
+        test_login_id = '004551ca-21fa-4dbe-9095-0384e73b3cbe'
+
+        application = models.Application.objects.create(
+            application_id=(UUID(test_application_id)),
+            application_type='CHILDMINDER',
+            application_status='DRAFTING',
+            cygnum_urn='',
+            login_details_status='COMPLETED',
+            personal_details_status='COMPLETED',
+            childcare_type_status='COMPLETED',
+            first_aid_training_status='COMPLETED',
+            eyfs_training_status='COMPLETED',
+            criminal_record_check_status='COMPLETED',
+            health_status='COMPLETED',
+            references_status='COMPLETED',
+            people_in_home_status='COMPLETED',
+            declarations_status='IN_PROGRESS',
+            date_created=datetime.datetime.today(),
+            date_updated=datetime.datetime.today(),
+            date_accepted=None,
+            order_code=None
+        )
+        user = models.UserDetails.objects.create(
+            application_id=application,
+            login_id=(UUID(test_login_id)),
+            email='',
+            mobile_number='',
+            add_phone_number='',
+            email_expiry_date=None,
+            sms_expiry_date=None,
+            magic_link_email='',
+            magic_link_sms=''
+        )
+        application_status_context = dict({
+            'application_id': test_application_id,
+            'login_details_status': application.login_details_status,
+            'personal_details_status': application.personal_details_status,
+            'childcare_type_status': application.childcare_type_status,
+            'first_aid_training_status': application.first_aid_training_status,
+            'eyfs_training_status': application.eyfs_training_status,
+            'criminal_record_check_status': application.criminal_record_check_status,
+            'health_status': application.health_status,
+            'reference_status': application.references_status,
+            'people_in_home_status': application.people_in_home_status,
+            'declaration_status': application.declarations_status,
+            'all_complete': True,
+            'confirm_details': False
+        })
+        assert (application_status_context['declaration_status'] == 'IN_PROGRESS')
+
+    def test_status_update_with_tasks_completed_with_declarations_complete(self):
+        test_application_id = 'f8c42666-1367-4878-92e2-1cee6ebcb48c'
+        test_login_id = '004551ca-21fa-4dbe-9095-0384e73b3cbe'
+
+        application = models.Application.objects.create(
+            application_id=(UUID(test_application_id)),
+            application_type='CHILDMINDER',
+            application_status='DRAFTING',
+            cygnum_urn='',
+            login_details_status='COMPLETED',
+            personal_details_status='COMPLETED',
+            childcare_type_status='COMPLETED',
+            first_aid_training_status='COMPLETED',
+            eyfs_training_status='COMPLETED',
+            criminal_record_check_status='COMPLETED',
+            health_status='COMPLETED',
+            references_status='COMPLETED',
+            people_in_home_status='COMPLETED',
+            declarations_status='COMPLETED',
+            date_created=datetime.datetime.today(),
+            date_updated=datetime.datetime.today(),
+            date_accepted=None,
+            order_code=None
+        )
+        user = models.UserDetails.objects.create(
+            application_id=application,
+            login_id=(UUID(test_login_id)),
+            email='',
+            mobile_number='',
+            add_phone_number='',
+            email_expiry_date=None,
+            sms_expiry_date=None,
+            magic_link_email='',
+            magic_link_sms=''
+        )
+        application_status_context = dict({
+            'application_id': test_application_id,
+            'login_details_status': application.login_details_status,
+            'personal_details_status': application.personal_details_status,
+            'childcare_type_status': application.childcare_type_status,
+            'first_aid_training_status': application.first_aid_training_status,
+            'eyfs_training_status': application.eyfs_training_status,
+            'criminal_record_check_status': application.criminal_record_check_status,
+            'health_status': application.health_status,
+            'reference_status': application.references_status,
+            'people_in_home_status': application.people_in_home_status,
+            'declaration_status': application.declarations_status,
+            'all_complete': True,
+            'confirm_details': True
+        })
+        assert (application_status_context['declaration_status'] == 'COMPLETED')
+
+    def delete(self):
+        models.Application.objects.get(pk='f8c42666-1367-4878-92e2-1cee6ebcb48c').delete()
+        models.UserDetails.objects.get(login_id='004551ca-21fa-4dbe-9095-0384e73b3cbe').delete()
