+++ application/tests/tests_views/test_validate_magic_link_email.py
@@ -1,0 +1,12 @@
+from .test_views import *
+
+
+class ValidateMagicLinkEmailTest(ViewsTest):
+
+    def test_page_not_displayed_without_magic_link_code(self):
+        c = Client()
+        try:
+            c.get(settings.URL_PREFIX + '/validate/')
+            self.assertEqual(1, 0)
+        except:
+            self.assertEqual(0, 0)
